+++ kitty/launch.py
@@ -18,7 +18,7 @@ child process.
 
 --tab-title
 The title for the new tab if launching in a new tab. By default, the title
-of the actie window in the tab is used as the tab title.
+of the active window in the tab is used as the tab title.
 
 
 --type
@@ -73,11 +73,12 @@ newly launched child process.
 --location
 type=choices
 default=last
-choices=first,neighbor,last
+choices=first,after,before,neighbor,last
 Where to place the newly created window when it is added to a tab which
-already has existing windows in it. Also applies to creating a new tab,
-where the value of neighbor will cause the new tab to be placed next to
-the current tab instead of at the end.
+already has existing windows in it. :code:`after` and :code:`before` place the new
+window before or after the active window. :code:`neighbor` is a synonym for :code:`after`.
+Also applies to creating a new tab, where the value of :code:`after`
+will cause the new tab to be placed next to the current tab instead of at the end.
 
 
 --allow-remote-control
@@ -116,6 +117,9 @@ want to pipe to program that wants to duplicate the screen layout of the
 screen.
 
 
+--marker
+Create a marker that highlights text in the newly created window. The syntax is
+the same as for the :code:`toggle_marker` map action (see :doc:`/marks`).
 '''
         options_spec.ans = OPTIONS
     return options_spec.ans
@@ -144,7 +148,7 @@ def get_env(opts, active_child):
 def tab_for_window(boss, opts, target_tab=None):
     if opts.type == 'tab':
         tm = boss.active_tab_manager
-        tab = tm.new_tab(empty_tab=True, as_neighbor=opts.location == 'neighbor')
+        tab = tm.new_tab(empty_tab=True, location=opts.location)
         if opts.tab_title:
             tab.set_title(opts.tab_title)
     elif opts.type == 'os-window':
@@ -178,6 +182,8 @@ def launch(boss, opts, args, target_tab=None):
         kw['override_title'] = opts.window_title
     if opts.copy_colors and active:
         kw['copy_colors_from'] = active
+    if opts.marker:
+        kw['marker'] = opts.marker
     cmd = args or None
     if opts.copy_cmdline and active_child:
         cmd = active_child.foreground_cmdline