+++ docs/changelog.rst
@@ -4,14 +4,20 @@ Changelog
 |kitty| is a feature full, cross-platform, *fast*, GPU based terminal emulator.
 To update |kitty|, :doc:`follow the instructions <binary>`.
 
-0.15.2 [future]
+0.16.0 [future]
 --------------------
 
+- A new :doc:`marks` feature that allows highlighting and scrolling to arbitrary
+  text in the terminal window.
+
 - hints kitten: Allow pressing :sc:`goto_file_line` to quickly open
-  the selected file at the selected line in vim (:iss:`2268`)
+  the selected file at the selected line in vim or a configurable editor (:iss:`2268`)
 
 - Allow choosing OpenType features for individual fonts via the
   :opt:`font_features` option.
+
+- Allow opening new tabs/windows before the current tab/window as well as after
+  it with the :option:`launch --location` option.
 
 - Add a :opt:`resize_in_steps` option that can be used to resize the OS window
   in steps as large as character cells (:pull:`2131`)